Title: 최솟값 3개, Time: 67ms, Memory: 1MB, Status: Passed - Codetree

+++ 260306/최솟값 3개/top-3-smallest-number.cpp
@@ -18,7 +18,7 @@ int main() {
         }
         else
         {   
-            int total = 1 ;
+            long long total = 1 ;
             int tmp[3] = {0,};
             for(int j = 0 ; j < 3; j++)
             {
